Fix dictionary tests importing from one another, fix not handling v2 block range type
